Use spaces, not tabs, for indentation.
There were a bunch of 4-space tab characters in the file, which is 1)
not the way UN*Xes work and 2) not what the modelines say; replace them
with 4 spaces, and further adjust some indentation.

Change the modelines to turn tab-to-space expansion on.

Change-Id: I7e22294e928ef95ab9f5d61f5d0e8abfe18cfb4e
Reviewed-on: https://code.wireshark.org/review/36738
Reviewed-by: Guy Harris <gharris@sonic.net>
(cherry picked from commit a543cdd3506e34859b10fd210db6c965446f56cb)
Reviewed-on: https://code.wireshark.org/review/36739