Add additional checks, clean up some stuff.
Add checks for bad block lengths - either too short or not a multiple of
4.  (Yes, the pcapng spec requires it to be a multiple of 4.  And there
is at least one implementation that requires it.)

For various structures with a length field, create the top-level tree
field for the item with a "run to the end of the packet" length and,
once we're finished dissecting it, set the length to its actual value.

Fetch various field values using proto_tree_item_add_uint.  Fix some
incorrect field types based on errors reported by that.

If an end-of-options option has a non-zero length, 1) don't treat it as
not an end-of-options option and 2) report an error on its length.

Change-Id: I72b2c065f3e3c76d5b71a1cd2ef3c1f497623266
Reviewed-on: https://code.wireshark.org/review/36746
Petri-Dish: Guy Harris <gharris@sonic.net>
Tested-by: Petri Dish Buildbot
Reviewed-by: Guy Harris <gharris@sonic.net>
(cherry picked from commit 0d2e6033ecad9c1e33239b5976a7c1b356fe4173)
Reviewed-on: https://code.wireshark.org/review/37118