Bluetooth: nordic: Add support for protocol version 3
Add protocol for the nRF Sniffer for BLE version 3 of the UART protocol.
These changes provides support for giving different packet IDs for advertising
physical channel and data physical channel.
The flags for advertising physical changes are intepreted differently,
the direction, encryption and MIC valid flags are always zero and are therefore
marked as reserved for future use instead.
The time_delta field is changed to be a firmware timestamp instead. This is to
allow better timestamping of the packets as the timestamp earlies was provided
by the extcap python code, which does not provide accurate timestamps.

Change-Id: I7f0ad74b3e30ffecade59b6d0c5965bfc634531c
Signed-off-by: Joakim Andersson <joakim.andersson@nordicsemi.no>
Reviewed-on: https://code.wireshark.org/review/36786
Petri-Dish: Stig Bjørlykke <stig@bjorlykke.org>
Tested-by: Petri Dish Buildbot
Reviewed-by: Anders Broman <a.broman58@gmail.com>