Bluetooth: Add support for dissecting packet format on LE Coded PHY
Add support for dissecting packet format on LE Coded PHY. This only includes
one additional field, the Coding Indicator (CI) which provides information
about the symbol rate of the FEC Block 2 of the pdu.
The TERM1 and TERM2 bytes are like the preamble assumed not included in the PDU,
these blocks are just bit-sequences for the radio and contains no important
information.

Change-Id: I7f0ad74b3e30ffecade59b6d0c5965bfc634531d
Signed-off-by: Joakim Andersson <joakim.andersson@nordicsemi.no>
Reviewed-on: https://code.wireshark.org/review/36787
Petri-Dish: Anders Broman <a.broman58@gmail.com>
Tested-by: Petri Dish Buildbot
Reviewed-by: Anders Broman <a.broman58@gmail.com>