Add record length checks.
Make sure the summary record is large enough; if not, report it as a bad
file.

If it's *too* large, skip the added data.

Clean up the length check for the header records - use sizeof, as we
later use sizeof when subtracting the fixed length portion's length.

Change-Id: I70697804eaa0cbbb1fb074eadf6457d237f26876
Reviewed-on: https://code.wireshark.org/review/36814
Petri-Dish: Guy Harris <gharris@sonic.net>
Tested-by: Petri Dish Buildbot
Reviewed-by: Guy Harris <gharris@sonic.net>