Fix display of header length and timestamps in UFTP4
The headers for UFTP version 4 contain a header length field which gives
the length of the header in 4 byte words.  Currently, only the raw value
is displayed, not the actual byte count (for example 4 instead of 16).

Several headers contain a timestamp field composed of 4 byte seconds and
4 byte microseconds since the UNIX epoch.  These are currently being
interpreted incorrectly as nanoseconds instead of microseconds.

The FILEINFO header contains a file timestamp field composed of 4 bytes
seconds since the epoch that is currently displayed as a raw value
instead of as a timestamp.

Change-Id: I936eb5317ca6802a094d8c1e01ae8ae78bb5cb7c
Reviewed-on: https://code.wireshark.org/review/36910
Petri-Dish: Jaap Keuter <jaap.keuter@xs4all.nl>
Tested-by: Petri Dish Buildbot
Reviewed-by: Jaap Keuter <jaap.keuter@xs4all.nl>