pcapng: show some fields in decimal, not hexadecimal.
The interface ID is just an ordinal; there's no reason to show it as hex
(we don't show it as hex if we're treating a pcapng file as a capture
rather than a file to be dissected).

The packet drops count is just a count, so, again, there's no reason to
show it as hex.

The hash algorithms numbers are given in decimal in the pcapng spec, so
display it as decimal.

Change-Id: I93fd50e7243a5b012bd29324f7116e634aca62af
Reviewed-on: https://code.wireshark.org/review/37072
Petri-Dish: Guy Harris <gharris@sonic.net>
Tested-by: Petri Dish Buildbot
Reviewed-by: Guy Harris <gharris@sonic.net>
(cherry picked from commit ab00f3533cd96fa1ff8f691be07ebb01b6f43e2c)
Reviewed-on: https://code.wireshark.org/review/37079