packet-kerberos: maintain a kerberos_all_keys map
This will be used temporary in order to replace
enc_key_list until we maintain/use separate lists.

Note that this will use wmem_file_scope(), but it
will get the longterm keys filled in when needed.

In the long run, we'll use more detailed lists
and use optimizations depending on the key usage.

Change-Id: If654dcfbc9ec8742eadbbb82b97a23fe8403022d
Signed-off-by: Stefan Metzmacher <metze@samba.org>
Reviewed-on: https://code.wireshark.org/review/37287
Petri-Dish: Anders Broman <a.broman58@gmail.com>
Tested-by: Petri Dish Buildbot
Reviewed-by: Anders Broman <a.broman58@gmail.com>