version_info: make indentation more consistent.
The preprocessor statements in the function get_compiler_info were
indented with a mixed approach: indented, but not following the style
of the rest of the file. The indented approach has been followed to
increase the readability, but with tabs, as in the rest of the file.

Other indentation issues fixed.

Change-Id: I6486c2cfa640f87c90a7fe4b7d95848d02249d23
Reviewed-on: https://code.wireshark.org/review/37481
Reviewed-by: Alexis La Goutte <alexis.lagoutte@gmail.com>
Petri-Dish: Guy Harris <gharris@sonic.net>
Tested-by: Petri Dish Buildbot
Reviewed-by: Anders Broman <a.broman58@gmail.com>