acdr: cleaner test for "is this tree item an RTP item?"
Don't rely on the item having the name "Real-Time Transport Protocol",
as that's fragile.  We could check for the abbreviation or filter name
being "rtp", but we can avoid string comparisons entirely by getting the
ID for the protocol "rtp" and comparing the item's ID with that.

Change-Id: I2a859f0b346f15a797fb5de75e0cf5c0b5abcfe9
Reviewed-on: https://code.wireshark.org/review/37490
Petri-Dish: Guy Harris <gharris@sonic.net>
Reviewed-by: Orgad Shaneh <orgads@gmail.com>
Reviewed-by: Anders Broman <a.broman58@gmail.com>